Correct calculation of total lessons to provide accurate course progress
Cast lesson length to integer to fix type mismatch in total lessons calculation.

+++ flutter_app/lib/services/course_service.dart
@@ -198,7 +198,7 @@ class CourseService {
       // Calculate total number of lessons in the course
       int totalLessons = 0;
       for (final module in course.modules ?? []) {
-        totalLessons += module.lessons.length;
+        totalLessons += module.lessons.length.toInt();
       }
       
       // Calculate progress percentage
